[#7] Move some logic inside service;
